Update tuklib_integer.h
A new version of the Intel compiler (not yet released) requires that _bit_scan_reverse be used via #include <immintrin.h> ; this is because it will be implemented via an inline assembler call.  If you use the intrinsic without the #include then at link time there is an undefined symbol [c code], or if in c++ code, there is a compilation error due to undefined name.   I work for Intel on the Intel C++ compiler team. Melanie Blower melanie dot blower at intel dot com